Moved StockTools and fixed imports.
