+++ ts/Stock/StockTools/StockToolsDefaults.ts
